Added Writing Data to Excel workbook
Added code to write trip data to an existing Excel workbook table.
Added methods to TripInfo to convert start and end times to DateTime format in both GMT and LocalTime.
Added aggregating the fuel consumption for the data base on consumption of each leg and consumption between beginning of the day and the start of the first leg and the end of the last leg and the end of the day.